// src/app/models/admin/dangkyhoatdongkhoangsan/dkdongcuamo.model.ts
export class OutputDkDongCuaMoModel {
  iddangkykhaithac: string;
  dientichdongcua: number;
  dientichkhaithac: number;
  donvidientich: string;
  geom: string;
  geovn: string;
  geowgs: string;
  hequychieu: string;
  idhoso: string;
  idgiayphep: string;
  sogiayphep: string;
  ngaycapphep: string;
  loaidongcua: number;
  siteid: string;
}

export class InputDkDongCuaMoModel {
  iddangkykhaithac: string;
  dientichdongcua: number;
  dientichkhaithac: number;
  donvidientich: string;
  hequychieu: string;
  idhoso: string;
  idgiayphep: string;
  loaidongcua: number;
  siteid: string;
}